Refactor loan account and order handling; remove logger and enhance transfer functionality
- Updated loan-account.ts to utilize akToAccountIdCache for fetching loan account ID.
- Removed logger.ts as it was deemed unnecessary.
- Refactored market-order.ts to streamline imports and improve code clarity.
- Introduced order-actions.ts to handle order submission, modification, and cancellation with improved structure.
- Updated order.ts to replace accountUid$ with akToAccountIdCache for fetching trading account ID.
- Refactored services.ts to use akToAccountIdCache for trading account ID retrieval.
- Enhanced strategy-account.ts to utilize akToAccountIdCache for strategy account ID.
- Refactored trade.ts to improve account trade retrieval logic and utilize client cache.
- Added transfer.ts to manage asset transfers with detailed handling for various account types and conditions.
- Updated pnpm-lock.yaml to include new dependencies and ensure proper linking.
