+++ apps/vendor-okx/package.json
@@ -19,6 +19,8 @@
     "@yuants/data-order": "workspace:*",
     "@yuants/utils": "workspace:*",
     "@yuants/sql": "workspace:*",
+    "@yuants/cache": "workspace:*",
+    "@yuants/secret": "workspace:*",
     "@yuants/data-series": "workspace:*",
     "@yuants/data-product": "workspace:*",
     "@yuants/data-ohlc": "workspace:*",
@@ -26,7 +28,6 @@
     "@yuants/data-quote": "workspace:*",
     "@yuants/data-trade": "workspace:*",
     "rxjs": "~7.5.6",
-    "crypto-js": "^4.2.0",
     "ws": "~8.8.1"
   },
   "devDependencies": {
