+++ apps/vendor-okx/src/transfer.ts
@@ -1,0 +1,294 @@
+import { createCache } from '@yuants/cache';
+import { Terminal } from '@yuants/protocol';
+import { addAccountTransferAddress } from '@yuants/transfer';
+import { formatTime } from '@yuants/utils';
+import { accountConfigCache } from './account';
+import { client$, clientCache } from './api';
+
+const terminal = Terminal.fromNodeEnv();
+
+const resOfAssetCurrenciesCache = createCache(
+  async (ak: string) => {
+    const client = await clientCache.query(ak);
+    if (!client) throw new Error('Client not found');
+    return client.getAssetCurrencies();
+  },
+  {
+    expire: 3600_000,
+  },
+);
+
+client$.subscribe(async (client) => {
+  const access_key = client.auth.access_key;
+  const account_config = await accountConfigCache.query(access_key).then((x) => x!.data[0]);
+  console.info(formatTime(Date.now()), 'AccountConfig', JSON.stringify(account_config));
+  const { mainUid, uid } = account_config;
+  const isMainAccount = mainUid === uid;
+
+  const TRADING_ACCOUNT_ID = `okx/${uid}/trading`;
+  const FUNDING_ACCOUNT_ID = `okx/${uid}/funding/USDT`;
+  const EARNING_ACCOUNT_ID = `okx/${uid}/earning/USDT`;
+
+  // BLOCK_CHAIN: only available for main account
+  if (isMainAccount) {
+    const depositAddressRes = await client.getAssetDepositAddress({ ccy: 'USDT' });
+    console.info(formatTime(Date.now()), 'DepositAddress', JSON.stringify(depositAddressRes.data));
+    const addresses = depositAddressRes.data.filter((v) => v.chain === 'USDT-TRC20' && v.to === '6');
+    for (const address of addresses) {
+      addAccountTransferAddress({
+        terminal,
+        account_id: FUNDING_ACCOUNT_ID,
+        network_id: 'TRC20',
+        currency: 'USDT',
+        address: address.addr,
+        onApply: {
+          INIT: async (order) => {
+            if (
+              !order.current_amount ||
+              order.current_amount < 3 // 最小提币额度
+            ) {
+              return { state: 'ERROR', message: 'Amount too small' };
+            }
+            const res = await resOfAssetCurrenciesCache.query(access_key);
+            const theRes = res?.data?.find((x) => x.ccy === 'USDT' && x.chain === 'USDT-TRC20');
+            const _fee = theRes?.minFee;
+            if (!_fee) return { state: 'ERROR', message: 'Currency Info not found, cannot get fee' };
+            const fee = +_fee;
+            const amt = Math.floor(order.current_amount - fee);
+            const transferResult = await client.postAssetWithdrawal({
+              amt: `${amt}`,
+              ccy: 'USDT',
+              chain: 'USDT-TRC20',
+              fee: `${fee}`,
+              dest: '4', // 链上提币
+              toAddr: order.current_rx_address!,
+            });
+            if (transferResult.code !== '0') {
+              return { state: 'INIT', message: transferResult.msg };
+            }
+            const wdId = transferResult.data[0]?.wdId;
+            return { state: 'AWAIT_TX_ID', context: wdId };
+          },
+          AWAIT_TX_ID: async (transferOrder) => {
+            const wdId = transferOrder.current_tx_context;
+            const withdrawalHistory = await client.getAssetWithdrawalHistory({ wdId });
+            const txId = withdrawalHistory.data?.[0]?.txId;
+            if (!txId) {
+              return { state: 'AWAIT_TX_ID', context: wdId };
+            }
+            return { state: 'COMPLETE', transaction_id: txId };
+          },
+        },
+        onEval: async (transferOrder) => {
+          const checkResult = await client.getAssetDepositHistory({
+            ccy: 'USDT',
+            txId: transferOrder.current_transaction_id,
+            type: '4',
+          });
+
+          if (checkResult.code !== '0') {
+            return {
+              state: 'INIT',
+            };
+          }
+
+          if (checkResult.data[0].state !== '2') {
+            return { state: 'PENDING' };
+          }
+          const received_amount = +checkResult.data[0].amt;
+          return { state: 'COMPLETE', received_amount };
+        },
+      });
+    }
+  }
+
+  // Funding-Trading
+  {
+    const FUNDING_TRADING_NETWORK_ID = `OKX/${uid}/Funding-Trading`;
+    addAccountTransferAddress({
+      terminal,
+      account_id: FUNDING_ACCOUNT_ID,
+      network_id: FUNDING_TRADING_NETWORK_ID,
+      currency: 'USDT',
+      address: 'funding',
+      onApply: {
+        INIT: async (order) => {
+          const transferResult = await client.postAssetTransfer({
+            type: '0',
+            ccy: 'USDT',
+            amt: `${order.current_amount}`,
+            from: '6',
+            to: '18',
+          });
+          if (transferResult.code !== '0') {
+            return { state: 'INIT', message: transferResult.msg };
+          }
+          const transaction_id = transferResult.data[0].transId;
+          return { state: 'COMPLETE', transaction_id };
+        },
+      },
+      onEval: async (transferOrder) => {
+        return { state: 'COMPLETE', received_amount: transferOrder.current_amount };
+      },
+    });
+    addAccountTransferAddress({
+      terminal,
+      account_id: TRADING_ACCOUNT_ID,
+      network_id: FUNDING_TRADING_NETWORK_ID,
+      currency: 'USDT',
+      address: 'trading',
+      onApply: {
+        INIT: async (order) => {
+          const transferResult = await client.postAssetTransfer({
+            type: '0',
+            ccy: order.currency,
+            amt: `${order.current_amount}`,
+            from: '18',
+            to: '6',
+          });
+          if (transferResult.code !== '0') {
+            return { state: 'INIT', message: transferResult.msg };
+          }
+          const transaction_id = transferResult.data[0].transId;
+          return { state: 'COMPLETE', transaction_id };
+        },
+      },
+      onEval: async (transferOrder) => {
+        return { state: 'COMPLETE', received_amount: transferOrder.current_amount };
+      },
+    });
+  }
+  // Funding-Earning
+  {
+    const FUNDING_EARNING_NETWORK_ID = `OKX/${uid}/Funding-Earning`;
+    addAccountTransferAddress({
+      terminal,
+      account_id: FUNDING_ACCOUNT_ID,
+      network_id: FUNDING_EARNING_NETWORK_ID,
+      currency: 'USDT',
+      address: 'funding',
+      onApply: {
+        INIT: async (order) => {
+          const transferResult = await client.postFinanceSavingsPurchaseRedempt({
+            ccy: 'USDT',
+            amt: `${order.current_amount}`,
+            side: 'purchase',
+            rate: '0.01',
+          });
+          if (transferResult.code !== '0') {
+            return { state: 'INIT', message: transferResult.msg };
+          }
+          return { state: 'COMPLETE', transaction_id: 'ok' };
+        },
+      },
+      onEval: async (transferOrder) => {
+        return { state: 'COMPLETE', received_amount: transferOrder.current_amount };
+      },
+    });
+    addAccountTransferAddress({
+      terminal,
+      account_id: EARNING_ACCOUNT_ID,
+      network_id: FUNDING_EARNING_NETWORK_ID,
+      currency: 'USDT',
+      address: 'earning',
+      onApply: {
+        INIT: async (order) => {
+          const transferResult = await client.postFinanceSavingsPurchaseRedempt({
+            ccy: 'USDT',
+            amt: `${order.current_amount}`,
+            side: 'redempt',
+            rate: '0.01',
+          });
+          if (transferResult.code !== '0') {
+            return { state: 'INIT', message: transferResult.msg };
+          }
+          return { state: 'COMPLETE', transaction_id: 'ok' };
+        },
+      },
+      onEval: async (transferOrder) => {
+        return { state: 'COMPLETE', received_amount: transferOrder.current_amount };
+      },
+    });
+  }
+
+  // SubAccount
+  {
+    const getSubAccountNetworkId = (subUid: string) => `OKX/${mainUid}/SubAccount/${subUid}`;
+    if (isMainAccount) {
+      const subAcctsRes = await client.getSubAccountList();
+      for (const item of subAcctsRes.data || []) {
+        addAccountTransferAddress({
+          terminal,
+          account_id: FUNDING_ACCOUNT_ID,
+          network_id: getSubAccountNetworkId(item.uid),
+          currency: 'USDT',
+          address: 'main',
+          onApply: {
+            INIT: async (order) => {
+              const transferResult = await client.postAssetTransfer({
+                type: '1',
+                ccy: 'USDT',
+                amt: `${order.current_amount}`,
+                from: '6',
+                to: '6',
+                subAcct: item.subAcct,
+              });
+              if (transferResult.code !== '0') {
+                return { state: 'INIT', message: transferResult.msg };
+              }
+              const transaction_id = transferResult.data[0].transId;
+              return { state: 'COMPLETE', transaction_id };
+            },
+          },
+          onEval: async (order) => {
+            // ISSUE: OKX API Issue: transId is incorrect or transId does not match with ‘ type’
+            // const checkResult = await client.getAssetTransferState({ transId: order.current_transaction_id });
+            // const received_amount = checkResult?.data?.[0]?.amt;
+            // if (!received_amount) {
+            //   return { state: 'INIT', message: checkResult.msg };
+            // }
+            // return { state: 'COMPLETE', received_amount: +received_amount };
+
+            return { state: 'COMPLETE', received_amount: order.current_amount };
+          },
+        });
+      }
+    }
+    // SubAccount API
+    else {
+      addAccountTransferAddress({
+        terminal,
+        account_id: FUNDING_ACCOUNT_ID,
+        network_id: getSubAccountNetworkId(uid),
+        currency: 'USDT',
+        address: 'sub',
+        onApply: {
+          INIT: async (order) => {
+            const transferResult = await client.postAssetTransfer({
+              type: '3',
+              ccy: 'USDT',
+              amt: `${order.current_amount}`,
+              from: '6',
+              to: '6',
+            });
+            if (transferResult.code !== '0') {
+              return { state: 'INIT', message: transferResult.msg };
+            }
+            const transaction_id = transferResult.data[0].transId;
+            return { state: 'COMPLETE', transaction_id };
+          },
+        },
+        onEval: async (order) => {
+          // ISSUE: OKX API Issue: transId is incorrect or transId does not match with ‘ type’
+          // const checkResult = await client.getAssetTransferState({ transId: order.current_transaction_id });
+          // const received_amount = checkResult?.data?.[0]?.amt;
+          // if (!received_amount) {
+          //   return { state: 'INIT', message: checkResult.msg };
+          // }
+          // return { state: 'COMPLETE', received_amount: +received_amount };
+          return { state: 'COMPLETE', received_amount: order.current_amount };
+        },
+      });
+    }
+  }
+});
